Create sorted numbers
It basically lets you insert an array in any order and sorts it for you from the lowest number to the highest number, while it does that it also deals with the second parameter and puts that as it is going through the first array. Then it returns the index where (num) was inserted.
The last if/else statement is just in case it runs into another number already in the array that is the same as (num), it returns the index of the first instance of it.